+++ addons/main/CfgVehicles.hpp
@@ -31,6 +31,49 @@ class CfgVehicles {
         tf_range = 40000;
     };
 
+    class TFAR_rt1523g: TFAR_Bag_Base {
+        mass = 70;
+        maximumLoad = 80;
+    };
+    class TFAR_rt1523g_big: TFAR_rt1523g {
+        mass = 100;
+        maximumLoad = 160;
+    };
+    class TFAR_rt1523g_sage: TFAR_rt1523g {
+        mass = 85;
+        maximumLoad = 120;
+    };
+
+    class TFAR_mr3000: TFAR_Bag_Base {
+        mass = 100;
+        maximumLoad = 160;
+        tf_range = 40000;
+    };
+
+    class TFAR_anprc155: TFAR_Bag_Base {
+        mass = 100;
+        maximumLoad = 160;
+        tf_range = 40000;
+    };
+    class TFAR_anarc164: TFAR_Bag_Base {
+        mass = 85;
+        maximumLoad = 120;
+    };
+    class TFAR_anarc210: TFAR_Bag_Base {
+        mass = 85;
+        maximumLoad = 120;
+    };
+    class TFAR_mr6000l: TFAR_Bag_Base {
+        mass = 85;
+        maximumLoad = 120;
+    };
+
+    class TFAR_bussole: TFAR_Bag_Base {
+        mass = 70;
+        maximumLoad = 80;
+        tf_range = 40000;
+    };
+
     class LandVehicle;
     class Tank: LandVehicle {
         class ACE_SelfActions {};