+++ mod.cpp
@@ -1,6 +1,5 @@
-name = "Weapons balance - Vanilla - TFAR compat";
+name = "Weapons Balance - TFAR";
 description = "A collection of weight balancing mods made by johnb43";
-tooltip = "Weight balance";
-tooltipOwned = "Weight balance";
-overview = "";
+tooltip = "Weapons Balance - TFAR";
+tooltipOwned = "Weapons Balance - TFAR";
 author = "johnb43";